Moved all utility function source files from `utils/` to `src/`
