+++ src/networkUtils.mts
